requirements.txt
airtable-python-wrapper==0.15.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
numpy==1.19.3
pandas==1.1.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.11
